requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-environ==0.4.5
django-filter==2.2.0
django-search-views==0.3.7
djangorestframework==3.11.0
envs==1.3
gunicorn==20.0.4
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.2.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
uuid==1.30
validators==0.14.2
whois==0.9.6
